requirements.txt
customtkinter
pymupdf
openpyxl
pillow
tkinterdnd2
